@@ -2,18 +2,23 @@
 
 namespace App\Controller\Api;
 
+use App\Context\AppContext;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 class ExternalAuthController extends AbstractController
 {
-    /**
-     * @Route("/auth", name="app_request_auth")
-     */
-    public function requestAuthenticationAction(): Response
+    #[Route(path: '/auth', name: 'app_request_auth')]
+    public function requestAuthenticationAction(AppContext $appContext, string $appHost): Response
     {
-        return new Response();
+        return new Response(
+            $this->renderView('security/unauthorized_access.html.twig', [
+                'isPairing' => (bool) $appContext->getServer()?->isPairing(),
+                'device' => $appContext->getConnectedDevice(),
+                'appHost' => $appHost,
+            ]),
+            Response::HTTP_UNAUTHORIZED
+        );
     }
 }
