@@ -4,35 +4,27 @@ namespace App\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
 
-/**
- * @ORM\Entity(repositoryClass="App\Repository\HostRepository")
- *
- * @ORM\Table(name="host")
- *
- * @ORM\HasLifecycleCallbacks()
- */
+
+#[ORM\Table(name: 'host')]
+#[ORM\Entity(repositoryClass: 'App\Repository\HostRepository')]
+#[ORM\HasLifecycleCallbacks]
 class Host
 {
     use TimeableTrait;
 
-    /**
-     * @ORM\Id
-     *
-     * @ORM\GeneratedValue
-     *
-     * @ORM\Column(type="integer")
-     */
+    
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: 'integer')]
     protected ?int $id;
 
-    /**
-     * @ORM\Column(type="string", unique=true)
-     */
+    #[ORM\Column(type: 'string', unique: true)]
     private ?string $domain;
 
-    /** @ORM\ManyToOne(targetEntity="Server", inversedBy="host") */
+    #[ORM\ManyToOne(targetEntity: 'Server', inversedBy: 'host')]
     private ?Server $server = null;
 
-    /** @ORM\ManyToOne(targetEntity="Application", inversedBy="host") */
+    #[ORM\ManyToOne(targetEntity: 'Application', inversedBy: 'host')]
     private ?Application $app = null;
 
     public function __construct()
