@@ -78,7 +78,9 @@ class AddAdminCommand extends Command
         if (null !== $username) {
             $this->io->text(' > <info>Full Name</info>: '.$username);
         } else {
-            $username = $this->io->ask('Full Name', null, [$this->validator, 'validateUsername']);
+            $username = $this->io->ask('Full Name', null, function (?string $username) : string {
+                return $this->validator->validateUsername($username);
+            });
             $input->setArgument('username', $username);
         }
 
@@ -87,7 +89,9 @@ class AddAdminCommand extends Command
         if (null !== $email) {
             $this->io->text(' > <info>Email</info>: '.$email);
         } else {
-            $email = $this->io->ask('Email', null, [$this->validator, 'validateEmail']);
+            $email = $this->io->ask('Email', null, function (?string $email) : string {
+                return $this->validator->validateEmail($email);
+            });
             $input->setArgument('email', $email);
         }
 
@@ -98,7 +102,9 @@ class AddAdminCommand extends Command
         } else {
             $password = $this->io->askHidden(
                 'Password - min 6 chars (your type will be hidden)',
-                [$this->validator, 'validatePassword']
+                function (?string $plainPassword) : string {
+                    return $this->validator->validatePassword($plainPassword);
+                }
             );
             $input->setArgument('password', $password);
         }
