@@ -14,9 +14,7 @@ use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Exception\AuthenticationException;
 use Symfony\Component\Security\Core\Exception\CustomUserMessageAuthenticationException;
 use Symfony\Component\Security\Http\Authenticator\AbstractAuthenticator;
-use Symfony\Component\Security\Http\Authenticator\Passport\Badge\UserBadge;
 use Symfony\Component\Security\Http\Authenticator\Passport\Passport;
-use Symfony\Component\Security\Http\Authenticator\Passport\SelfValidatingPassport;
 
 class ForwardedAuthAuthenticator extends AbstractAuthenticator
 {
@@ -58,7 +56,7 @@ class ForwardedAuthAuthenticator extends AbstractAuthenticator
         $this->appContext->setCreateTrustedCookie(true);
 
         throw new CustomUserMessageAuthenticationException(
-            sprintf('[REQUEST AUTH] User unauthorized to access to the endpoint')
+            '[REQUEST AUTH] User unauthorized to access to the endpoint'
         );
     }
 
